requirements.txt
yfinance
python-dotenv
requests
google-generativeai